[Update] App refactor and modifications
