+++ app/server/api.py
@@ -1,10 +1,15 @@
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 from app.server.auth.authenticate import router as auth_router
-from app.server.policies import router as api_router
-from app.server.user import router as user_router
+from app.server.routes.policies import router as api_router
+from app.server.routes.user import router as user_router
+from app.server.routes.data import router as data_router
 
-app = FastAPI(swagger_ui_parameters={"syntaxHighlight.theme": "obsidian"})
+app = FastAPI(
+    swagger_ui_parameters={"syntaxHighlight.theme": "obsidian"},
+    title="Policy Management API",
+    version="0.1.0",
+)
 
 
 origins = ["*"]
@@ -21,3 +26,4 @@ app.add_middleware(
 app.include_router(auth_router)
 app.include_router(api_router)
 app.include_router(user_router)
+app.include_router(data_router)
