+++ app/server/auth/authenticate.py
@@ -3,26 +3,11 @@ import json
 from fastapi import APIRouter
 import requests as rest_client
 from fastapi import APIRouter
-from starlette.responses import RedirectResponse
-from starlette.requests import Request
 
-router = APIRouter()
+router = APIRouter(tags=["Token"])
 
 
-from app.config.config import settings
-
-
-@router.get("/gitlab/login")
-async def login_gitlab():
-    APP_ID, REDIRECT_URI = (
-        settings.GITLAB_CLIENT_ID,
-        "http://localhost:8080/api/auth/callback/gitlab",
-    )
-    url = f"https://gitlab.com/oauth/authorize?client_id={APP_ID}&redirect_uri={REDIRECT_URI}&response_type=code"
-    return RedirectResponse(url=url)
-
-
-@router.get("/api/auth/callback/gitlab")
+@router.get("/gitlab/token")
 async def get_token_from_gitlab(
     code: str, client_id: str, client_secret: str, redirect_uri: str
 ) -> dict:
@@ -34,7 +19,6 @@ async def get_token_from_gitlab(
             "code": code,
             "grant_type": "authorization_code",
             "redirect_uri": redirect_uri,
-            # "redirect_uri": "http://localhost:8080/api/auth/callback/gitlab",
         },
     )
     json_res = res.json()
@@ -42,7 +26,7 @@ async def get_token_from_gitlab(
     return {"access_token": access_token, "expires_in": expires_in}
 
 
-@router.get("/token")
+@router.get("/github/token")
 def get_token_from_github(code: str, client_id: str, client_secret: str) -> dict:
     res = rest_client.post(
         url="https://github.com/login/oauth/access_token",
